Rewrite underlying data structure
With the previous data structure, cart data was spread across many Redis
keys. This made it possible to query Redis for an answer without needing
to load the entire data set. But for most cart applications, you need to
ask many queries on your cart data during one page request. This leads
to lots of chatter back and forth between your app server, and Redis
instance.

This rewrite focuses on putting all the cart data in one JSON encoded
Redis string key. This allows the cart to be loaded once per request,
and many queries to be run on the data in memory without the overhead of
talking to Redis each time.

This version also includes a conversion script that will automatically
migrate cart data to the new format the first time it is requested.